requirements.txt
torch
torchvision 
numpy 
pyyaml 
Pillow 
torchview
graphviz
